Repeat first milestone of admin page with form to create a new competition
This last change was to replicate the previous ability to create a
page of competitions, plus a blank form to create a new one.  But this
uses javascript to load the content of the page, as a proof of concept
to using asynchronous calls to load some of the content.