Fix validation of integers problem
Integer values should validate as OK when there is nothing in the
field yet (maybe this is an empty field.  This was not happening. In
fact the opposite was the case, probably because of a change made
earlier when I switched the signs of this function and then changed my
mind again and must have just switched part of it back