Output version of mbball as part of version string
Since we have problems with the browser cashing mbbball.js, it will be
useful to know the version of the script it is working with.  Add in
the version of mbball as part of the version string.

(Note this also implies that you do not have to up the version of
index if we are only changing the javascript.