This new branch adds the ability to dump the postgress database to a series of sqlite ones.
A new function dump.php, to be called in a browser from the applications
baseurl/dump.php will dump the existing postgress database and produce
a series of sqlite databases.

football.ini is an overall configuration database which says which is
the default competition, the remainder are databases which each have a
single competition in them.

This is so that sqlite version of this application can use them