Change auth mechanism to use a remote call to an auth site to fetch the important stuff
This is a new mechanism for authorising the user remotely.  A new php
function see remote/football.php sits at the authorising site and can
check out what ever mechanism is needed - in this case smf login
account, but it doesn't have to be that.  The required information is
returned as javascript code to write a cookie and redirect back to the
starting point.

All this is encrypted to prevent evesdropping for account data - and
the calling parameters are also encrypted to prevent eavesdroppers
replicating the call and obtaining data that way.